removed email from existing code and moved style.css to one folder.adjusted sinup.js but still having errors.modified homeroutes to try asnd render homepage without login
