--- public/style.css
@@ -1,64 +1,0 @@
-body {
-    font-family: Arial, sans-serif;
-    margin: 0;
-    padding: 0;
-    background-color: #e0f7fa;
-}
-
-header {
-    background-color: #0288d1;
-    color: white;
-    padding: 10px 0;
-}
-
-h1 {
-    margin: 0;
-    padding: 0 20px;
-    float: left;
-}
-
-nav {
-    float: right;
-    padding-right: 20px;
-}
-
-nav a {
-    color: white;
-    text-decoration: none;
-    margin: 0 10px;
-}
-
-.container {
-    width: 80%;
-    margin: auto;
-    overflow: hidden;
-}
-
-.container-fluid {
-    padding: 20px;
-}
-
-.no-button {
-    background: none;
-    color: inherit;
-    border: none;
-    padding: 0;
-    font: inherit;
-    cursor: pointer;
-    outline: inherit;
-}
-
-footer {
-    background-color: #0288d1;
-    color: white;
-    text-align: center;
-    padding: 10px;
-    position: fixed;
-    bottom: 0;
-    width: 100%;
-}
-
-
-
-
-
