+++ views/homepage.handlebars
@@ -1,0 +1,14 @@
+<div class="row">
+  <div class="col-md-6">
+    <h2>Login</h2>
+    <form class="form login-form">
+    </form>
+    <script src="/login.js"></script>
+  </div>
+  <div class="col-md-6">
+    <h2>Sign Up</h2>
+    <form class="form signup-form">
+    </form>
+    <script src="/signup.js"></script>
+  </div>
+</div>
