+++ views/layouts/main.handlebars
@@ -6,7 +6,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Project X!</title>
   <link rel="stylesheet" href="/css/jass.css">
-  <link rel="stylesheet" href="/style.css">
+  <link rel="stylesheet" href="/css/style.css">
 </head>
 
 <body>
