+++ views/signup.handlebars
@@ -7,10 +7,6 @@
         <input class="form-input" type="text" id="username-signup" name="username" />
       </div>
       <div class="form-group">
-        <label for="email-signup">Email:</label>
-        <input class="form-input" type="email" id="email-signup" name="email" />
-      </div>
-      <div class="form-group">
         <label for="password-signup">Password:</label>
         <input class="form-input" type="password" id="password-signup" name="password" />
       </div>
@@ -22,4 +18,4 @@
   <div class="col-md-6">
   </div>
 </div>
-<script src="/"></script>
+<script src="/js/signup.js"></script>
